requirements.txt
fastapi
uvicorn
jinja2
instaloader
python-multipart
sse-starlette
transformers
torch
torchvision
pillow